gui/requirements.txt
customtkinter
pygame-ce
pyinstaller